refactor: change addRowHoverClassByPosition  to setHoveredRowKeyByPosition and apply it
