+++ packages/toast-ui.grid/src/dispatch/data.ts
@@ -8,7 +8,6 @@ import {
   Row,
   Column,
   Range,
-  PagePosition,
   LoadingState,
   PageOptions,
   ColumnInfo
@@ -63,9 +62,6 @@ import {
   updateAllSummaryValues
 } from './summary';
 import { initFilter } from './filter';
-import { cls } from '../helper/dom';
-import { setHoveredRowKey } from './renderState';
-import { findRowIndexByPosition } from '../query/mouse';
 import { OriginData } from './types';
 import { getSelectionRange } from '../query/selection';
 import { initScrollPosition } from './viewport';
@@ -678,26 +674,6 @@ export function removeRowClassName(store: Store, rowKey: RowKey, className: stri
   if (row) {
     removeArrayItem(className, row._attributes.className.row);
     notify(row._attributes, 'className');
-  }
-}
-
-export function addRowHoverClassByPosition(store: Store, viewInfo: PagePosition) {
-  const {
-    renderState: { hoveredRowKey },
-    data: { filteredRawData },
-    viewport: { scrollLeft, scrollTop }
-  } = store;
-  const rowIndex = findRowIndexByPosition(store, {
-    ...viewInfo,
-    scrollLeft,
-    scrollTop
-  });
-  const rowKey = filteredRawData[rowIndex].rowKey;
-
-  if (hoveredRowKey !== rowKey) {
-    removeRowClassName(store, hoveredRowKey!, cls('row-hover'));
-    setHoveredRowKey(store, rowKey);
-    addRowClassName(store, rowKey, cls('row-hover'));
   }
 }
